First pass at creating a land module
