+++ docs/modules.rst
@@ -1,0 +1,27 @@
+Modules
+=======
+
+Atmos
+-----
+.. automodule:: xclim.atmos._precip
+   :members:
+   :private-members:
+   :inherited-members:
+   :show-inheritance:
+   :undoc-members:
+
+.. automodule:: xclim.atmos._temperature
+   :members:
+   :private-members:
+   :inherited-members:
+   :show-inheritance:
+   :undoc-members:
+
+Land
+----
+.. automodule:: xclim.land
+   :members:
+   :private-members:
+   :inherited-members:
+   :show-inheritance:
+   :undoc-members:
